IGAPP-403: Move more builders and hashUrl.js to api-client
