+++ api-client/package.json
@@ -18,8 +18,9 @@
   "dependencies": {
     "@babel/polyfill": "^7.8.3",
     "@babel/runtime": "^7.8.3",
+    "js-md5": "^0.7.3",
     "entities": "^2.0.0",
-    "moment": "^2.24",
+    "moment": "2.25.3",
     "moment-timezone": "^0.5",
     "normalize-path": "^3.0.0",
     "sanitize-html-react": "^1.13"
@@ -28,6 +29,7 @@
     "@babel/core": "^7.8.3",
     "@babel/preset-env": "^7.8.3",
     "@babel/preset-flow": "^7.8.3",
+    "seedrandom": "^3.0.3",
     "jest": "^25.5.4",
     "jest-fetch-mock": "^3.0.3",
     "jest-junit": "^12.0.0"
