+++ native/package.json
@@ -51,7 +51,6 @@
     "htmlparser2": "^4.0.0",
     "i18next": "^19.1.0",
     "is-url": "^1.2.4",
-    "js-md5": "^0.7.3",
     "translations": "0.0.1",
     "lodash": "^4.17.19",
     "moment": "2.25.3",
@@ -125,7 +124,6 @@
     "redux-flipper": "^1.4.2",
     "redux-mock-store": "^1.5.4",
     "redux-saga-test-plan": "^4.0.0-rc.1",
-    "seedrandom": "^3.0.3",
     "wait-for-expect": "^3.0.2",
     "wd": "^1.12.1",
     "yo": "^3.1.1"
