+++ native/src/modules/categories/components/__tests__/Categories.spec.js
@@ -6,7 +6,7 @@ import CityModelBuilder from 'api-client/src/testing/CityModelBuilder'
 import LanguageModelBuilder from 'api-client/src/testing/LanguageModelBuilder'
 import Page from '../../../common/components/Page'
 import createNavigationScreenPropMock from '../../../../testing/createNavigationStackPropMock'
-import CategoriesMapModelBuilder from '../../../../testing/builder/CategoriesMapModelBuilder'
+import CategoriesMapModelBuilder from 'api-client/src/testing/CategoriesMapModelBuilder'
 import Categories from '../Categories'
 import lightTheme from '../../../theme/constants'
 import CategoriesRouteStateView from '../../../app/CategoriesRouteStateView'
