+++ native/src/modules/endpoint/__tests__/DefaultDataContainer.spec.js
@@ -6,8 +6,8 @@ import RNFetchBlob from '../../../__mocks__/rn-fetch-blob'
 import moment from 'moment'
 import CityModelBuilder from 'api-client/src/testing/CityModelBuilder'
 import LanguageModelBuilder from 'api-client/src/testing/LanguageModelBuilder'
-import CategoriesMapModelBuilder from '../../../testing/builder/CategoriesMapModelBuilder'
-import EventModelBuilder from '../../../testing/builder/EventModelBuilder'
+import CategoriesMapModelBuilder from 'api-client/src/testing/CategoriesMapModelBuilder'
+import EventModelBuilder from 'api-client/src/testing/EventModelBuilder'
 import DatabaseConnector from '../DatabaseConnector'
 import PoiModelBuilder from 'api-client/src/testing/PoiModelBuilder'
