+++ native/src/modules/endpoint/sagas/__tests__/watchContentLanguageSwitch.spec.js
@@ -6,11 +6,11 @@ import type { SwitchContentLanguageActionType } from '../../../app/StoreActionTy
 import LanguageModelBuilder from 'api-client/src/testing/LanguageModelBuilder'
 import { expectSaga, testSaga } from 'redux-saga-test-plan'
 import loadCityContent from '../loadCityContent'
-import CategoriesMapModelBuilder from '../../../../testing/builder/CategoriesMapModelBuilder'
+import CategoriesMapModelBuilder from 'api-client/src/testing/CategoriesMapModelBuilder'
 import watchContentLanguageSwitch, { switchContentLanguage } from '../watchContentLanguageSwitch'
 import AsyncStorage from '@react-native-community/async-storage'
 import AppSettings from '../../../settings/AppSettings'
-import EventModelBuilder from '../../../../testing/builder/EventModelBuilder'
+import EventModelBuilder from 'api-client/src/testing/EventModelBuilder'
 import PoiModelBuilder from 'api-client/src/testing/PoiModelBuilder'
 
 jest.mock('rn-fetch-blob')
