+++ native/src/modules/endpoint/sagas/__tests__/watchFetchCategory.spec.js
@@ -7,7 +7,7 @@ import LanguageModelBuilder from 'api-client/src/testing/LanguageModelBuilder'
 import watchFetchCategory, { fetchCategory } from '../watchFetchCategory'
 import { expectSaga, testSaga } from 'redux-saga-test-plan'
 import loadCityContent from '../loadCityContent'
-import CategoriesMapModelBuilder from '../../../../testing/builder/CategoriesMapModelBuilder'
+import CategoriesMapModelBuilder from 'api-client/src/testing/CategoriesMapModelBuilder'
 import ErrorCodes from '../../../error/ErrorCodes'
 import moment from 'moment'
 import mockDate from '../../../../testing/mockDate'
