+++ native/src/modules/endpoint/sagas/__tests__/watchFetchEvent.spec.js
@@ -3,7 +3,7 @@
 import RNFetchBlob from '../../../../__mocks__/rn-fetch-blob'
 import DefaultDataContainer from '../../DefaultDataContainer'
 import type { FetchEventActionType } from '../../../app/StoreActionType'
-import EventModelBuilder from '../../../../testing/builder/EventModelBuilder'
+import EventModelBuilder from 'api-client/src/testing/EventModelBuilder'
 import LanguageModelBuilder from 'api-client/src/testing/LanguageModelBuilder'
 import watchFetchEvent, { fetchEvent } from '../watchFetchEvent'
 import { expectSaga, testSaga } from 'redux-saga-test-plan'
