+++ native/src/modules/endpoint/sagas/fetchResourceCache.js
@@ -11,7 +11,7 @@ import moment from 'moment'
 import type { LanguageResourceCacheStateType, PageResourceCacheStateType } from '../../app/StateType'
 import { fromError } from '../../error/ErrorCodes'
 
-export type FetchMapTargetType = { url: string, filePath: string, urlHash: string }
+export type FetchMapTargetType = {| url: string, filePath: string, urlHash: string |}
 export type FetchMapType = { [path: string]: Array<FetchMapTargetType> }
 
 const createErrorMessage = (fetchResult: FetchResultType) => {
