+++ native/src/routes/events/components/__tests__/Events.spec.js
@@ -4,7 +4,7 @@ import * as React from 'react'
 import TestRenderer from 'react-test-renderer'
 import lightTheme from '../../../../modules/theme/constants'
 import Events from '../Events'
-import EventModelBuilder from '../../../../testing/builder/EventModelBuilder'
+import EventModelBuilder from 'api-client/src/testing/EventModelBuilder'
 import CityModelBuilder from 'api-client/src/testing/CityModelBuilder'
 import LanguageModelBuilder from 'api-client/src/testing/LanguageModelBuilder'
 import Page from '../../../../modules/common/components/Page'
